Extract ProgressSteps component
